[Refactor] Share on-policy trainer implementation across PPO/A2C/REINFORCE
Extract the trainer body previously triplicated across PPOTrainer,
A2CTrainer and ReinforceTrainer into a common OnPolicyTrainer base
class; the algorithm classes now only override the algorithm name and
the default number of epochs (4 for PPO, 1 for A2C/REINFORCE). The same
consolidation is applied to the Hydra layer with OnPolicyTrainerConfig
and a shared _make_onpolicy_trainer factory.

This fixes several issues in the A2C/REINFORCE additions:

- Default GAE hyperparameters were gamma=0.9, lmbda=0.99 (copied from a
  stale docstring); they now inherit PPO's gamma=0.99, lmbda=0.95.
- A2CTrainerConfig/ReinforceTrainerConfig were missing most trainer
  kwargs (gamma, lmbda, enable_logging, log_*, key remapping, hooks,
  batch_size); they now expose the full field set, and the factories
  restore the _normalize_hydra_key and _register_trainer_hooks handling
  as well as the nn.Module guard before instantiating networks.
- A new lr_scheduler kwarg/config field wires LRSchedulerHook into all
  three trainers (accepting a partial config taking the optimizer).
- GAE is registered through ValueEstimatorHook (checkpointable, and
  safe under async collection) instead of a bare register_op, and the
  A2C/REINFORCE trainers gain PPO's done/reward/action key plumbing.
- _make_a2c_loss/_make_reinforce_loss no longer forward gamma,
  advantage_key and value_target_key to constructors that reject them;
  they are routed through make_value_estimator and set_keys.
- batch_size, which was silently ignored, now emits a warning pointing
  at the replay buffer's batch_size.
- The sota-implementations entry points import the single config class
  they need instead of a wildcard import.
