+++ torchrl/trainers/algorithms/configs/__init__.py
@@ -129,6 +129,7 @@ from torchrl.trainers.algorithms.configs.trainers import (
     DQNTrainerConfig,
     IQLTrainerConfig,
     OfflineToOnlineTrainerConfig,
+    OnPolicyTrainerConfig,
     PPOTrainerConfig,
     ReinforceTrainerConfig,
     SACTrainerConfig,
@@ -410,6 +411,7 @@ __all__ = [
     "DQNTrainerConfig",
     "IQLTrainerConfig",
     "OfflineToOnlineTrainerConfig",
+    "OnPolicyTrainerConfig",
     "PPOTrainerConfig",
     "ReinforceTrainerConfig",
     "SACTrainerConfig",
